pyproject.toml
[tool.poetry]
name = "cicero"
version = "0.1.0"
description = "Team Apex's entry for AI JSON 2024 Summer hackathon"
authors = ["Raphael Darley <raphael@raphaeldarley.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
aijson-meta = "^0.1.3"
fastapi = "^0.112.0"
fastapi-cli = "^0.0.5"
litellm = "1.42.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
